Tworzenie nowej kampanii działa

+++ AdvertApi/AdvertApi/AdvertApi/Models/Banner.cs
@@ -10,9 +10,9 @@ namespace AdvertApi.Models
     {
         public int IdAdvertisment { get; set; }
         public int Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int? IdCampaign { get; set; }
         public Campaign Campaign { get; set; }
-        public double Area { get; set; }
+        public decimal Area { get; set; }
     }
 }
