+++ AdvertApi/AdvertApi/AdvertApi/Models/Building.cs
@@ -13,7 +13,7 @@ namespace AdvertApi.Model
         public string Street { get; set; }
         public int StreetNumber { get; set; }
         public string City { get; set; }
-        public double Height { get; set; }
+        public decimal Height { get; set; }
         public virtual ICollection<Campaign> CampaignsFrom { get; set; }
         public virtual ICollection<Campaign> CampaignsTo { get; set; }
